server/requirements.txt
blinker==1.8.2
branca==0.8.1
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.7
colorama==0.4.6
contourpy==1.3.1
cycler==0.12.1
dnspython==2.7.0
Flask==3.0.3
Flask-Cors==5.0.0
folium==0.19.4
fonttools==4.55.4
geopandas==1.0.1
greenlet==3.1.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.0
networkx==3.4.2
numpy==2.2.2
osmnx==1.3.0
packaging==24.2
pandas==2.2.3
pillow==11.0.0
pyogrio==0.10.0
pyparsing==3.2.1
pyproj==3.7.0
python-dateutil==2.9.0.post0
pytz==2024.2
requests==2.32.3
scikit-learn==1.5.0
scipy==1.15.1
shapely==2.0.6
six==1.17.0
SQLAlchemy==2.0.36
threadpoolctl==3.5.0
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
Werkzeug==3.0.6
xyzservices==2025.1.0
